Added Rose Gold Armor
Rose Gold Armor is meant to be the stepping stone between Iron Armor and Diamond Armor. It has the properties of Iron Armor, however the enchantable level of golden armor. It also has durability between iron and diamond armor. It is crafted using rose gold ingots (4 copper and 2 gold in any pattern).